+++ src/main/java/com/example/chatserver/entity/ChatMemberSetting.java
@@ -29,5 +29,5 @@ public class ChatMemberSetting {
 
     @Column(name="created_date")
     @Nonnull
-    private Integer createdDate;
+    private Long createdDate;
 }
